+++ src/frameworks/angular/index.ts
@@ -25,7 +25,6 @@ import {
   getBuildConfig,
   getContext,
   getServerConfig,
-  maybeWarnAngular22SsrSecurity,
 } from "./utils";
 import { I18N_ROOT, SHARP_VERSION } from "../constants";
 import { FirebaseError } from "../../error";
@@ -69,10 +68,8 @@ export async function build(dir: string, configuration: string): Promise<BuildRe
     locales,
     baseHref: baseUrl,
     ssr,
-    buildTargetOptions,
   } = await getBuildConfig(dir, configuration);
   await warnIfCustomBuildScript(dir, name, DEFAULT_BUILD_SCRIPT);
-  await maybeWarnAngular22SsrSecurity(dir, { ssr, buildTargetOptions });
   for (const target of targets) {
     // TODO there is a bug here. Spawn for now.
     // await scheduleTarget(prerenderTarget);